requirements.txt
pandas 
gspread 
google-auth